+++ splinter/driver/webdriver/chrome.py
@@ -2,6 +2,7 @@
 
 from selenium.webdriver import Chrome
 from splinter.driver.webdriver import BaseWebDriver, WebDriverElement
+from splinter.driver.webdriver.cookie_manager import CookieManager
 
 
 class WebDriver(BaseWebDriver):
@@ -11,6 +12,8 @@ class WebDriver(BaseWebDriver):
         self._unpatch_subprocess()
 
         self.element_class = WebDriverElement
+
+        self._cookie_manager = CookieManager(self.driver)
 
         super(WebDriver, self).__init__()