+++ splinter/driver/zopetestbrowser.py
@@ -1,17 +1,45 @@
-import mimetypes
-
-import lxml.html
 from lxml.cssselect import CSSSelector
 from zope.testbrowser.browser import Browser
-
 from splinter.element_list import ElementList
 from splinter.driver import DriverAPI, ElementAPI
 from splinter.utils import warn_deprecated
+from splinter.cookie_manager import CookieManagerAPI
+
+import mimetypes
+import lxml.html
+
+
+class CookieManager(CookieManagerAPI):
+
+    def __init__(self, browser_cookies):
+        self._cookies = browser_cookies
+
+    def add(self, cookies):
+        for key, value in cookies.items():
+            self._cookies[key] = value
+
+    def delete(self, cookie=None):
+        if cookie:
+            try:
+                del self._cookies[cookie]
+            except KeyError:
+                pass
+        else:
+            self._cookies.clearAll()
+
+    def __getitem__(self, item):
+        return self._cookies[item]
+
+    def __eq__(self, other_object):
+        if isinstance(other_object, dict):
+            return dict(self._cookies) == other_object
+
 
 class ZopeTestBrowser(DriverAPI):
 
     def __init__(self):
         self._browser = Browser()
+        self._cookie_manager = CookieManager(self._browser.cookies)
         self._last_urls = []
 
     def visit(self, url):
@@ -131,13 +159,18 @@ class ZopeTestBrowser(DriverAPI):
         return ElementList([ZopeTestBrowserLinkElement(link, self) for link in links])
 
     def select(self, name, value):
-        self.find_by_name(name).first._control.value = [value,]
+        self.find_by_name(name).first._control.value = [value]
 
     def _element_is_link(self, element):
         return element.tag == 'a'
 
     def _element_is_control(self, element):
         return hasattr(element, 'type')
+
+    @property
+    def cookies(self):
+        return self._cookie_manager
+
 
 class ZopeTestBrowserElement(ElementAPI):
 
@@ -148,9 +181,33 @@ class ZopeTestBrowserElement(ElementAPI):
     def __getitem__(self, attr):
         return self._element.attrib[attr]
 
+    def find_by_css(self, selector):
+        elements = self._element.cssselect(selector)
+        return ElementList([self.__class__(element, self) for element in elements])
+
+    def find_by_xpath(self, selector):
+        elements = self._element.xpath(selector)
+        return ElementList([self.__class__(element, self) for element in elements])
+
+    def find_by_name(self, name):
+        elements = self._element.cssselect('[name="%s"]' % name)
+        return ElementList([self.__class__(element, self) for element in elements])
+
+    def find_by_tag(self, name):
+        elements = self._element.cssselect(name)
+        return ElementList([self.__class__(element, self) for element in elements])
+
+    def find_by_id(self, id):
+        elements = self._element.cssselect('#%s' % id)
+        return ElementList([self.__class__(element, self) for element in elements])
+
     @property
     def value(self):
         return self._element.text
+
+    @property
+    def text(self):
+        return self.value
 
 
 class ZopeTestBrowserLinkElement(ZopeTestBrowserElement):
@@ -167,6 +224,7 @@ class ZopeTestBrowserLinkElement(ZopeTestBrowserElement):
 
     def click(self):
         return self._browser.open(self["href"])
+
 
 class ZopeTestBrowserControlElement(ElementAPI):
 
@@ -187,6 +245,7 @@ class ZopeTestBrowserControlElement(ElementAPI):
 
     def click(self):
         return self._control.click()
+
 
 class ZopeTestBrowserOptionElement(ElementAPI):