+++ docs/news.rst
@@ -6,6 +6,21 @@
 splinter news
 +++++++++++++
 
+what's new in splinter 0.2?
+===========================
+
+Features
+--------
+
+- :doc:`cookies manipulation </cookies>`
+- find elements within an element
+- improvements in `ElementList`
+
+Backward incompatible changes
+-----------------------------
+
+- you should update your selenium to 2.1.0 version and your chrome driver. See more in :doc:`suport to new chrome driver </drivers/chrome>` 
+
 what's new in splinter 0.1.1?
 =============================